pipeline/requirements.txt
geopandas>=0.14
shapely>=2.0
fiona>=1.9
pdfplumber>=0.11
requests>=2.31
beautifulsoup4>=4.12
psycopg2-binary>=2.9
supabase>=2.0
python-dotenv>=1.0
